+++ src/api/config.ts
@@ -1,0 +1,2 @@
+export { PIPELINE_VERSION } from '../modules/pipeline/config'
+export { VALID_TIMEFRAMES, MAX_CANDLE_LIMIT } from '../modules/binance/constants'